fix: move `AllowedFieldTypes` and `AllowedQuestionTypes`
