+++ src/argilla/client/feedback/dataset/base.py
@@ -28,8 +28,8 @@ from argilla.client.feedback.schemas import (
     RankingQuestion,
     RatingQuestion,
 )
+from argilla.client.feedback.schemas.types import AllowedFieldTypes, AllowedQuestionTypes
 from argilla.client.feedback.training.schemas import TrainingTaskMappingForTextClassification
-from argilla.client.feedback.types import AllowedFieldTypes, AllowedQuestionTypes
 from argilla.client.feedback.unification import (
     LabelQuestionStrategy,
     MultiLabelQuestionStrategy,
